refactor(coherence): split tests into sibling modules for near-cap files
fleet.rs and routes_parse.rs were within 22 lines of the 300-line
per-file cap and mixed production code with their test fixtures,
leaving no headroom for the next feature. Move the unit tests into
sibling files, matching the existing adrs_tests / agents_tests /
handshake_tests pattern (ADR-0040,  13).CONSTITUTION

  fleet.rs           287 -> 166 lines
  routes_parse.rs    278 -> 238 lines

Two private helpers (fleet::build_report, routes_parse::extract_methods
and parse_doc_methods) are bumped to pub(crate) so the sibling
test modules can call them.

Also restore the plan_execution unit tests (infer_type, required_kinds)
that were accidentally dropped in the preceding fix commit. They now
live alongside the new build_report mock test in plan_execution_tests.rs.
