+++ crates/convergio-coherence/src/routes_parse.rs
@@ -86,7 +86,7 @@ pub(super) fn parse_route_lines(body: &str, file: &str) -> Vec<RouteEntry> {
     out
 }
 
-fn extract_methods(rest: &str) -> BTreeSet<String> {
+pub(crate) fn extract_methods(rest: &str) -> BTreeSet<String> {
     const VERBS: &[&str] = &["get", "post", "put", "patch", "delete", "head", "options"];
     let mut out: BTreeSet<String> = BTreeSet::new();
     for verb in VERBS {
@@ -199,7 +199,7 @@ fn extract_backtick_tokens(body: &str) -> Vec<String> {
     out
 }
 
-fn parse_doc_methods(cell: &str) -> BTreeSet<String> {
+pub(crate) fn parse_doc_methods(cell: &str) -> BTreeSet<String> {
     const VERBS: &[&str] = &["GET", "POST", "PUT", "PATCH", "DELETE", "HEAD", "OPTIONS"];
     let upper = cell.to_ascii_uppercase();
     let mut out: BTreeSet<String> = BTreeSet::new();
@@ -234,45 +234,4 @@ fn extract_doc_paths(cell: &str) -> Vec<String> {
 fn relativize(path: &Path) -> String {
     let s = path.to_string_lossy().to_string();
     s.trim_start_matches("./").to_string()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn extracts_methods_from_chain() {
-        let m = extract_methods(", get(handler).post(other))");
-        assert!(m.contains("GET"));
-        assert!(m.contains("POST"));
-    }
-
-    #[test]
-    fn parse_route_lines_basic() {
-        let body = r#"
-        Router::new()
-            .route("/v1/health", get(health))
-            .route("/v1/plans", post(create).get(list))
-            .route("/v1/internal", get(internal)) // docs-skip
-        "#;
-        let r = parse_route_lines(body, "routes/x.rs");
-        assert_eq!(r.len(), 2);
-        let plans = r.iter().find(|e| e.path == "/v1/plans").expect("plans");
-        assert!(plans.methods.contains("POST"));
-        assert!(plans.methods.contains("GET"));
-    }
-
-    #[test]
-    fn doc_methods_handles_dot_separator() {
-        let m = parse_doc_methods("POST · GET");
-        assert!(m.contains("POST"));
-        assert!(m.contains("GET"));
-    }
-
-    #[test]
-    fn docs_skip_exempts_route() {
-        let body = "        .route(\"/v1/internal/debug\", get(debug)) // docs-skip\n";
-        let r = parse_route_lines(body, "routes/x.rs");
-        assert!(r.is_empty());
-    }
 }
